mobile/android/tests/browser/chrome/test_debugger_server.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1010750
Migrated from Robocop: https://bugzilla.mozilla.org/show_bug.cgi?id=1184186
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1010750</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript;version=1.7">

  "use strict";

  const { utils: Cu } = Components;

  const DEBUGGER_USB_ENABLED = "devtools.remote.usb.enabled";

  Cu.import("resource://gre/modules/Services.jsm");
  const { require } =
    Cu.import("resource://gre/modules/devtools/shared/Loader.jsm", {});
  const { DebuggerServer } = require("devtools/server/main");

  let win = Services.wm.getMostRecentWindow("navigator:browser");

  win.RemoteDebugger.init();

  SimpleTest.registerCleanupFunction(function() {
    Services.prefs.clearUserPref(DEBUGGER_USB_ENABLED);
  });

  // Enable the debugger via the pref it listens for
  Services.prefs.setBoolPref(DEBUGGER_USB_ENABLED, true);

  ok(DebuggerServer.initialized, "initialized");
  is(DebuggerServer.listeningSockets, 1, "1 listening socket");

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1010750">Mozilla Bug 1010750</a>
<br>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1184186">Migrated from Robocop testDebuggerServer</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
